DisplayPackage perf fix
